pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "pimage"
version = "1.1.0"
keywords = ["image", "copy-move", "attack", "detection"]
authors = [
  { name="Rahmat Nazali Salimi", email="rahmatnazali95@gmail.com" },
]
maintainers = [
  { name="Rahmat Nazali Salimi", email="rahmatnazali95@gmail.com" },
]
description = "Library for detecting image copy-move attack"
readme = "README-PYPI.md"
requires-python = ">=3.8.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering :: Image Processing"
]
dependencies = [
  "numpy==1.24.1",
  "scikit-learn==1.2.1",
  "tqdm==4.64.1",
  "Pillow==9.4.0",
  "imageio==2.25.0",
]

[project.urls]
"Homepage" = "https://github.com/rahmatnazali/pimage"
"Bug Tracker" = "https://github.com/rahmatnazali/pimage/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["pimage"]
exclude = [
    ".github",
    "assets",
    "dataset",
    "example",
    "output",
    "venv"
]
namespaces = false